requirements.txt
certifi==2017.11.5
chardet==3.0.4
docker==2.7.0
docker-pycreds==0.2.1
gitdb2==2.0.3
GitPython==2.1.8
idna==2.6
pulsar==1.6.4
requests==2.18.4
six==1.11.0
smmap2==2.0.3
urllib3==1.22
websocket-client==0.45.0
